Pulse/requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
bs4==0.0.2
Django==5.2.10
django-filter==25.2
django-simple-history==3.11.0
djangorestframework==3.16.1
drf-nested-routers==0.95.0
Markdown==3.10.1
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
